pc - remove unnecessary files and refactor tests
