--- frontend/src/fixtures/temp.json
@@ -1,28 +1,0 @@
-[
-    {
-        "id": 1,
-        "email": "phtcon@ucsb.edu",
-        "googleSub": "115856948234298493496",
-        "pictureUrl": "https://lh3.googleusercontent.com/-bQynVrzVIrU/AAAAAAAAAAI/AAAAAAAAAAA/AMZuucmkGuVsELD1ZeV5iDUAUfe6_K-p8w/s96-c/photo.jpg",
-        "fullName": "Phill Conrad",
-        "givenName": "Phill",
-        "familyName": "Conrad",
-        "emailVerified": true,
-        "locale": "en",
-        "hostedDomain": "ucsb.edu",
-        "admin": true
-    },
-    {
-        "id": 2,
-        "email": "pconrad.cis@gmail.com",
-        "googleSub": "102656447703889917227",
-        "pictureUrl": "https://lh3.googleusercontent.com/a-/AOh14GhpDBUt8eCEqiRT45hrFbcimsX_h1ONn0dc3HV8Bp8=s96-c",
-        "fullName": "Phillip Conrad",
-        "givenName": "Phillip",
-        "familyName": "Conrad",
-        "emailVerified": true,
-        "locale": "en",
-        "hostedDomain": null,
-        "admin": false
-    }
-]
