--- frontend/src/fixtures/ucsbDatesFixtures.js
@@ -1,31 +1,0 @@
-const ucsbDatesFixtures = {
-    oneDate: {
-        "id": 1,
-        "quarterYYYYQ": "20221",
-        "name": "Noon on January 2nd",
-        "localDateTime": "2022-01-02T12:00:00"
-    },
-    threeDates: [
-        {
-            "id": 1,
-            "quarterYYYYQ": "20221",
-            "name": "Noon on January 2nd",
-            "localDateTime": "2022-01-02T12:00:00"
-        },
-        {
-            "id": 2,
-            "quarterYYYYQ": "20222",
-            "name": "Noon on April 3rd",
-            "localDateTime": "2022-04-03T12:00:00"
-        },
-        {
-            "id": 3,
-            "quarterYYYYQ": "20223",
-            "name": "Noon on July 4th",
-            "localDateTime": "2022-07-04T12:00:00"
-        }
-    ]
-};
-
-
-export { ucsbDatesFixtures };
