--- frontend/src/main/pages/Students/StudentsCreatePage.js
@@ -1,14 +1,0 @@
-import BasicLayout from "main/layouts/BasicLayout/BasicLayout";
-
-export default function StudentsCreatePage() {
-  return (
-    <BasicLayout>
-      <div className="pt-2">
-        <h1>Create New Student</h1>
-        <p>
-          This is where the create page will go
-        </p>
-      </div>
-    </BasicLayout>
-  )
-}
