--- frontend/src/main/pages/Todos/TodosEditPage.js
@@ -1,14 +1,0 @@
-import BasicLayout from "main/layouts/BasicLayout/BasicLayout";
-
-export default function TodosEditPage() {
-  return (
-    <BasicLayout>
-      <div className="pt-2">
-        <h1>Todos</h1>
-        <p>
-          This is where the edit page will go
-        </p>
-      </div>
-    </BasicLayout>
-  )
-}
