--- frontend/src/main/pages/Todos/TodosIndexPage.js
@@ -1,14 +1,0 @@
-import BasicLayout from "main/layouts/BasicLayout/BasicLayout";
-
-export default function TodosIndexPage() {
-  return (
-    <BasicLayout>
-      <div className="pt-2">
-        <h1>Todos</h1>
-        <p>
-          This is where the index page will go
-        </p>
-      </div>
-    </BasicLayout>
-  )
-}
