--- frontend/src/tests/pages/Students/StudentsCreatePage.test.js
@@ -1,31 +1,0 @@
-import { render } from "@testing-library/react";
-import StudentsCreatePage from "main/pages/Students/StudentsCreatePage";
-import { QueryClient, QueryClientProvider } from "react-query";
-import { MemoryRouter } from "react-router-dom";
-
-import { apiCurrentUserFixtures }  from "fixtures/currentUserFixtures";
-import { systemInfoFixtures } from "fixtures/systemInfoFixtures";
-import axios from "axios";
-import AxiosMockAdapter from "axios-mock-adapter";
-
-
-describe("StudentsCreatePage tests", () => {
-
-    const axiosMock =new AxiosMockAdapter(axios);
-    axiosMock.onGet("/api/currentUser").reply(200, apiCurrentUserFixtures.userOnly);
-    axiosMock.onGet("/api/systemInfo").reply(200, systemInfoFixtures.showingNeither);
-
-    const queryClient = new QueryClient();
-    test("renders without crashing", () => {
-        render(
-            <QueryClientProvider client={queryClient}>
-                <MemoryRouter>
-                    <StudentsCreatePage />
-                </MemoryRouter>
-            </QueryClientProvider>
-        );
-    });
-
-});
-
-
